move map function to utils
